New tests accomplished but Shims_mp still does not work within the main file

+++ NewShims/shims_mp.py
@@ -55,8 +55,6 @@ def getBestShims(pallet, items, k, solTorque, solItems, cfg, lock, surplus):
                 if vol > maxVol:
                     break
 
-
-
     # create the first shim
     sh = Shims(pallet, len(whip))
     Set = [sh]
@@ -158,16 +156,22 @@ def loadPallets(cfg):
         reader.close()    
     return pallets
         
-def fillPallet(pallet, items, limit, k, solTorque, solItems, cfg, lock):
+def fillPallet(pallet, items, k, solTorque, solItems, cfg, lock, limit):
     for item in items:
         if pallet.isFeasible(item, limit, k, solTorque, solItems, cfg, lock):
             pallet.putItem(item, solTorque, solItems, lock)
 
-def Solve(pallets, items, cfg, k, limit, secBreak): # items include kept on board
+def Solve(pallets, items, cfg, k, limit, secBreak, mode): # items include kept on board
 
-    print(f"\nParallel Shims for ACLP+RPDP")
+    if mode == "p":
+        print(f"\nParallel Shims for ACLP+RPDP")
+    else:
+        print(f"\nSerial Shims for ACLP+RPDP")
 
-    surplus = math.exp(-limit) + 0.9
+    # surplus = math.exp(-limit) + 0.9
+    # surplus = 1. + 2. * (1. - limit)
+    surplus = (2. - limit)
+
     print(f"surplus: {surplus:.2f}")
 
     numItems   = len(items)
@@ -177,40 +181,57 @@ def Solve(pallets, items, cfg, k, limit, secBreak): # items include kept on boar
     solTorque.value = 0.0
 
     solItems = mp.Array('i', range(numItems))
-    for i, _ in enumerate(solItems):
-        solItems[i] = -1 # not alocated to any pallet
-
-    procs = [None for _ in pallets]
+    for j, _ in enumerate(solItems):
+        solItems[j] = -1 # not alocated to any pallet
 
     lock  = mp.Lock()
+
+    for j, item in enumerate(items):
+        items[j].ID = j
+        if item.P > -1: # if alocated to a pallet
+            for i, p in enumerate(pallets):
+                if item.P == p.ID: # in case there is some consolidated among the items
+                    pallets[i].putItem(item, solTorque, solItems, lock)
+
+    # item.P = -1 if an item, -2 if a consollidated, or pallet ID.
+
+    procs = [None for _ in pallets] # each pallets has its own process
 
     # sort ascendent by CG distance
     pallets.sort(key=lambda x: abs(x.D), reverse=False)
 
-    for i, _ in enumerate(procs):
-        procs[i] = mp.Process( target=fillPallet, args=( pallets[i], items, limit, k, solTorque, solItems, cfg, lock) )
-        time.sleep(0.001)
-        procs[i].start()
-    
-    for proc in procs:
-        proc.join()
+    if mode == "p":
+        # parallel greedy phase
+        for i, _ in enumerate(procs):
+            procs[i] = mp.Process( target=fillPallet, args=( pallets[i], items, k, solTorque, solItems, cfg, lock, limit) )
+            time.sleep(0.001)
+            procs[i].start()
+        
+        for proc in procs:
+            proc.join()
 
-    for i, _ in enumerate(procs):
-        procs[i] = mp.Process( target=getBestShims, args=( pallets[i], items, k, solTorque, solItems, cfg, lock, surplus) )
-        procs[i].start()
-    
+        # parallel shims phase
+        for i, _ in enumerate(procs):
+            procs[i] = mp.Process( target=getBestShims, args=( pallets[i], items, k, solTorque, solItems, cfg, lock, surplus) )
+            procs[i].start()
+        
+        start = time.time()
+        while time.time() - start <= secBreak:
+            if not any(p.is_alive() for p in procs):
+                # All the processes are done, break now.
+                break
+        else:
+            # We only enter this if we didn't 'break' above.
+            print("timed out, killing all processes")
+            for p in procs:
+                p.terminate()
+                p.join()
 
-    start = time.time()
-    while time.time() - start <= secBreak:
-        if not any(p.is_alive() for p in procs):
-            # All the processes are done, break now.
-            break
-    else:
-        # We only enter this if we didn't 'break' above.
-        print("timed out, killing all processes")
-        for p in procs:
-            p.terminate()
-            p.join()
+    else: # serial
+        for i, _ in enumerate(pallets):
+            fillPallet(  pallets[i], items, k, solTorque, solItems, cfg, lock, limit) 
+            getBestShims(pallets[i], items, k, solTorque, solItems, cfg, lock, surplus)
+               
 
     # --- mount solution matrix
     Z = np.zeros((numPallets,numItems))
